implemented TransferDataSource with functions to save and update Transfer objects. Created corresponding use cases and integrated them into the ConfirmTransferViewModel.

+++ app/src/main/java/com/project/digitalbank/ui/features/deposit/DepositFormFragment.kt
@@ -73,7 +73,7 @@ class DepositFormFragment : Fragment() {
 
         if (value > 0f) {
             hideKeyboard()
-            val deposit = Deposit(value=value.toFloat())
+            val deposit = Deposit(value=value)
             saveDeposit(deposit)
         } else {
             showBottomSheet(message = "Insert a deposit value.")
